+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 
 - Rebuilt `DirectoryTree` with new `Tree` control.
+- Container widgets now have default height of `1fr`.
 - The default `width` of a `Label` is now `auto`.
 
 ### Fixed